fix: only delete user old avatar when it is different from new one

+++ back-end/src/main/java/com/huyphan/services/UserService.java
@@ -98,6 +98,7 @@ public class UserService implements UserDetailsService, MediatorComponent {
             throws UserException {
         User user = getCurrentUser();
         String updatedUsername = updateProfileData.getUsername();
+        String updatedAvatarUrl = updateProfileData.getAvatarUrl();
 
         if (!Objects.equals(user.getUsername(), updatedUsername) && userRepo.existsByUsername(
                 updatedUsername)) {
@@ -111,7 +112,7 @@ public class UserService implements UserDetailsService, MediatorComponent {
         user.setCountry(updateProfileData.getCountry());
         user.setPrivate(updateProfileData.isPrivate());
 
-        if (!Objects.equals(oldAvatarUrl, defaultAvatarUrl)) {
+        if (!updatedAvatarUrl.equals(oldAvatarUrl) && !Objects.equals(oldAvatarUrl, defaultAvatarUrl)) {
             awss3Util.deleteObject(oldAvatarUrl);
         }
